+++ src/DakotaIterator.hpp
@@ -225,12 +225,12 @@ public:
   virtual const IntResponseMap& all_responses() const;
 
   /// get the current number of samples
-  virtual int num_samples() const;
+  virtual size_t num_samples() const;
   /// reset sampling iterator to use at least min_samples
-  virtual void sampling_reset(int min_samples, bool all_data_flag, 
+  virtual void sampling_reset(size_t min_samples, bool all_data_flag, 
 			      bool stats_flag);
   /// set reference number of samples, which is a lower bound during reset 
-  virtual void sampling_reference(int samples_ref);
+  virtual void sampling_reference(size_t samples_ref);
 
   /// increment to next in sequence of refinement samples
   virtual void sampling_increment();
